JBPM-4859 - Edit FormService and ImageService tests closes #345
