+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-all/src/test/java/org/kie/server/integrationtests/jbpm/ImageServiceIntegrationTest.java
@@ -17,33 +17,20 @@ package org.kie.server.integrationtests.jbpm;
 
 import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.ClassRule;
 import org.junit.Test;
-import org.junit.rules.ExternalResource;
 import org.kie.server.api.model.KieContainerResource;
 import org.kie.server.api.model.ReleaseId;
-import org.kie.server.client.KieServicesClient;
 import org.kie.server.client.KieServicesException;
-import org.kie.server.client.ProcessServicesClient;
-import org.kie.server.client.UIServicesClient;
-import org.kie.server.integrationtests.shared.RestJmsSharedBaseIntegrationTest;
 
 import static org.junit.Assert.*;
 
-public class ImageServiceIntegrationTest extends RestJmsSharedBaseIntegrationTest {
+public class ImageServiceIntegrationTest extends JbpmKieServerBaseIntegrationTest {
 
     private static ReleaseId releaseId = new ReleaseId("org.kie.server.testing", "definition-project",
             "1.0.0.Final");
 
     private static final String CONTAINER_ID = "definition-project";
-
     private static final String HIRING_PROCESS_ID = "hiring";
-
-    private UIServicesClient uiServicesClient;
-    private ProcessServicesClient processClient;
-
-    @ClassRule
-    public static ExternalResource StaticResource = new DBExternalResource();
 
     @BeforeClass
     public static void buildAndDeployArtifacts() {
@@ -53,33 +40,15 @@ public class ImageServiceIntegrationTest extends RestJmsSharedBaseIntegrationTes
 
     }
 
-    @Override
-    protected KieServicesClient createDefaultClient() throws Exception {
-        KieServicesClient servicesClient = super.createDefaultClient();
-
-        uiServicesClient = servicesClient.getServicesClient(UIServicesClient.class);
-        processClient = servicesClient.getServicesClient(ProcessServicesClient.class);
-
-        return servicesClient;
-    }
-
-    @Before
-    public void cleanup() {
-        cleanupSingletonSessionId();
-
-    }
-
     @Test
     public void testGetProcessImageViaUIClientTest() throws Exception {
         KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
-
         String result = uiServicesClient.getProcessImage(CONTAINER_ID, HIRING_PROCESS_ID);
         logger.debug("Image content is '{}'", result);
         assertNotNull(result);
         assertFalse(result.isEmpty());
-
     }
 
     @Test(expected = KieServicesException.class)
@@ -88,7 +57,6 @@ public class ImageServiceIntegrationTest extends RestJmsSharedBaseIntegrationTes
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
         uiServicesClient.getProcessImage(CONTAINER_ID, "not-existing");
-
     }
 
     @Test
@@ -114,6 +82,5 @@ public class ImageServiceIntegrationTest extends RestJmsSharedBaseIntegrationTes
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
         uiServicesClient.getProcessInstanceImage(CONTAINER_ID, 9999l);
-
     }
 }
