+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-all/src/test/java/org/kie/server/integrationtests/jbpm/rest/FormServiceRestOnlyIntegrationTest.java
@@ -13,16 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.server.integrationtests.jbpm;
+package org.kie.server.integrationtests.jbpm.rest;
 
 import java.util.HashMap;
 import java.util.Map;
-import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import org.jboss.resteasy.client.ClientRequest;
 import org.jboss.resteasy.client.ClientResponse;
 import org.jboss.resteasy.client.ClientResponseFailure;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -33,23 +33,27 @@ import org.kie.server.api.marshalling.Marshaller;
 import org.kie.server.api.marshalling.MarshallerFactory;
 import org.kie.server.api.model.KieContainerResource;
 import org.kie.server.api.model.ReleaseId;
+import org.kie.server.api.model.instance.TaskSummary;
+import org.kie.server.api.model.instance.TaskSummaryList;
 import org.kie.server.api.model.type.JaxbLong;
 import org.kie.server.api.rest.RestURI;
-import org.kie.server.client.KieServicesClient;
 import org.kie.server.integrationtests.config.TestConfig;
+import org.kie.server.integrationtests.jbpm.DBExternalResource;
 import org.kie.server.integrationtests.shared.RestOnlyBaseIntegrationTest;
 
 import static org.junit.Assert.*;
 import static org.kie.server.api.rest.RestURI.*;
+import org.kie.server.client.KieServicesConfiguration;
 
-public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegrationTest{
+public class FormServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegrationTest {
 
     private static ReleaseId releaseId = new ReleaseId("org.kie.server.testing", "definition-project",
             "1.0.0.Final");
 
     private static final String CONTAINER_ID = "definition-project";
-
     private static final String HIRING_PROCESS_ID = "hiring";
+
+    private ClientResponse<?> response = null;
 
     @ClassRule
     public static ExternalResource StaticResource = new DBExternalResource();
@@ -63,9 +67,8 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
     }
 
     @Override
-    protected KieServicesClient createDefaultClient() {
-        KieServicesClient servicesClient = super.createDefaultClient();
-        return servicesClient;
+    protected void additionalConfiguration(KieServicesConfiguration configuration) {
+        configuration.setTimeout(30000);
     }
 
     @Before
@@ -74,8 +77,15 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
 
     }
 
+    @After
+    public void releaseConnection() {
+        if (response != null) {
+            response.releaseConnection();
+        }
+    }
+
     @Test
-    public void testGetProcessImageTest() throws Exception {
+    public void testGetProcessFormTest() throws Exception {
         KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
@@ -83,35 +93,27 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
         valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
         valuesMap.put(RestURI.PROCESS_ID, HIRING_PROCESS_ID);
 
-        ClientResponse<?> response = null;
         try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_IMG_GET_URI, valuesMap))
+            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), FORM_URI + "/" + PROCESS_FORM_GET_URI, valuesMap))
                     .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
+                    .header("Accept", getMediaType().toString());
+            logger.info("[GET] " + clientRequest.getUri());
 
             response = clientRequest.get();
             Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
 
             String result = response.getEntity(String.class);
-            logger.debug("Image content is '{}'", result);
+            logger.debug("Form content is '{}'", result);
             assertNotNull(result);
             assertFalse(result.isEmpty());
 
-
         } catch (Exception e) {
             throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
         }
-
     }
 
     @Test
-    public void testGetProcessInstanceImageTest() throws Exception {
+    public void testGetTaskFormTest() throws Exception {
         KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
@@ -121,9 +123,7 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
 
         Marshaller marshaller = MarshallerFactory.getMarshaller(marshallingFormat, ClassLoader.getSystemClassLoader());
 
-        ClientResponse<?> response = null;
         try {
-
             Map<String, Object> params = new HashMap<String, Object>();
             params.put("name", "john");
 
@@ -139,20 +139,40 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
             Long result = response.getEntity(JaxbLong.class).unwrap();
             assertNotNull(result);
 
+            // find tasks by process instance id
             valuesMap.put(RestURI.PROCESS_INST_ID, result);
 
-            clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_INST_IMG_GET_URI, valuesMap))
+            clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), QUERY_URI + "/" + TASK_BY_PROCESS_INST_ID_GET_URI, valuesMap))
                     .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
+                    .header("Accept", getMediaType().toString());
             logger.info("[GET] " + clientRequest.getUri());
 
             response = clientRequest.get();
             Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
 
-            String image = response.getEntity(String.class);
-            logger.debug("Image content is '{}'", image);
-            assertNotNull(image);
-            assertFalse(image.isEmpty());
+            TaskSummaryList taskSummaryList = marshaller.unmarshall(response.getEntity(String.class), TaskSummaryList.class);
+            logger.debug("Form content is '{}'", taskSummaryList);
+
+            assertNotNull(taskSummaryList);
+            TaskSummary[] task = taskSummaryList.getTasks();
+            assertEquals(1, task.length);
+
+            Long taskId = task[0].getId();
+
+            valuesMap.put(RestURI.TASK_INSTANCE_ID, taskId);
+
+            clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), FORM_URI + "/" + TASK_FORM_GET_URI, valuesMap))
+                    .header("Content-Type", getMediaType().toString())
+                    .header("Accept", getMediaType().toString());
+            logger.info("[GET] " + clientRequest.getUri());
+
+            response = clientRequest.get();
+            Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
+
+            String formdata = response.getEntity(String.class);
+            logger.debug("Form content is '{}'", formdata);
+            assertNotNull(formdata);
+            assertFalse(formdata.isEmpty());
 
             clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), PROCESS_URI + "/" + ABORT_PROCESS_INST_DEL_URI, valuesMap)).header("Content-Type", getMediaType().toString());
             logger.info("[DELETE] " + clientRequest.getUri());
@@ -161,15 +181,35 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
 
         } catch (Exception e) {
             throw new ClientResponseFailure(e, response);
-        } finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
         }
     }
 
     @Test
-    public void testGetProcessImageNotExistingTest() throws Exception {
+    public void testGetProcessDoesNotExistFormTest() throws Exception {
+        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
+        assertSuccess(client.createContainer(CONTAINER_ID, resource));
+
+        Map<String, Object> valuesMap = new HashMap<String, Object>();
+        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
+        valuesMap.put(RestURI.PROCESS_ID, "not-existing");
+        valuesMap.put(RestURI.TASK_INSTANCE_ID, 99999);
+
+        try {
+            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), FORM_URI + "/" + TASK_FORM_GET_URI, valuesMap))
+                    .header("Content-Type", getMediaType().toString())
+                    .header("Accept", getMediaType().toString());
+            logger.info("[GET] " + clientRequest.getUri());
+
+            response = clientRequest.get();
+            Assert.assertEquals(Response.Status.NOT_FOUND.getStatusCode(), response.getStatus());
+
+        } catch (Exception e) {
+            throw new ClientResponseFailure(e, response);
+        }
+    }
+
+    @Test
+    public void testGetTaskDoesNotExistFormTest() throws Exception {
         KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
         assertSuccess(client.createContainer(CONTAINER_ID, resource));
 
@@ -177,52 +217,17 @@ public class ImageServiceRestOnlyIntegrationTest extends RestOnlyBaseIntegration
         valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
         valuesMap.put(RestURI.PROCESS_ID, "not-existing");
 
-        ClientResponse<?> response = null;
         try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_IMG_GET_URI, valuesMap))
+            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), FORM_URI + "/" + PROCESS_FORM_GET_URI, valuesMap))
                     .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
+                    .header("Accept", getMediaType().toString());
+            logger.info("[GET] " + clientRequest.getUri());
 
             response = clientRequest.get();
             Assert.assertEquals(Response.Status.NOT_FOUND.getStatusCode(), response.getStatus());
 
         } catch (Exception e) {
             throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
-        }
-    }
-
-    @Test
-    public void testGetProcessInstanceImageNotExistingTest() throws Exception {
-        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
-        assertSuccess(client.createContainer(CONTAINER_ID, resource));
-
-        Map<String, Object> valuesMap = new HashMap<String, Object>();
-        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
-        valuesMap.put(RestURI.PROCESS_INST_ID, 9999);
-
-        ClientResponse<?> response = null;
-        try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_INST_IMG_GET_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
-
-            response = clientRequest.get();
-            Assert.assertEquals(Response.Status.NOT_FOUND.getStatusCode(), response.getStatus());
-
-        } catch (Exception e) {
-            throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
         }
     }
 }
